requirements.txt
numpy>=1.21.4
Pillow>=8.3.1
scipy>=1.7.2
torch>=1.10.0
